split files into directories
