+++ src/index.html
@@ -6,5 +6,6 @@
   </head>
   <body>
     <div id="game-grid"></div>
+    <div class="hidden"></div>
   </body>
 </html>
