+++ src/organisms/animals/Animal.js
@@ -1,10 +1,6 @@
-import { Animal } from './Animal';
+import { Organism } from '../Organism';
 
-class Antelope extends Animal {
-  divClass = 'antelope-tile';
-  initiative = 4;
-  strength = 4;
-
+class Animal extends Organism {
   constructor(xIndex, yIndex, board) {
     super(xIndex, yIndex, board);
   }
@@ -13,55 +9,55 @@ class Antelope extends Animal {
     if (randomNumber === 1) {
       return {
         xIndex: this.xIndex,
-        yIndex: this.yIndex - 2,
+        yIndex: this.yIndex - 1,
       };
     }
     if (randomNumber === 2) {
       return {
-        xIndex: this.xIndex + 2,
-        yIndex: this.yIndex - 2,
+        xIndex: this.xIndex + 1,
+        yIndex: this.yIndex - 1,
       };
     }
     if (randomNumber === 3) {
       return {
-        xIndex: this.xIndex + 2,
+        xIndex: this.xIndex + 1,
         yIndex: this.yIndex,
       };
     }
     if (randomNumber === 4) {
       return {
-        xIndex: this.xIndex + 2,
-        yIndex: this.yIndex + 2,
+        xIndex: this.xIndex + 1,
+        yIndex: this.yIndex + 1,
       };
     }
     if (randomNumber === 5) {
       return {
         xIndex: this.xIndex,
-        yIndex: this.yIndex + 2,
+        yIndex: this.yIndex + 1,
       };
     }
     if (randomNumber === 6) {
       return {
-        xIndex: this.xIndex - 2,
-        yIndex: this.yIndex + 2,
+        xIndex: this.xIndex - 1,
+        yIndex: this.yIndex + 1,
       };
     }
     if (randomNumber === 7) {
       return {
-        xIndex: this.xIndex - 2,
+        xIndex: this.xIndex - 1,
         yIndex: this.yIndex,
       };
     }
     if (randomNumber === 8) {
       return {
-        xIndex: this.xIndex - 2,
-        yIndex: this.yIndex - 2,
+        xIndex: this.xIndex - 1,
+        yIndex: this.yIndex - 1,
       };
     }
     if (randomNumber === 9) {
       return {
-        xIndex: this.xIndex - 2,
-        yIndex: this.yIndex - 2,
+        xIndex: this.xIndex - 1,
+        yIndex: this.yIndex - 1,
       };
     }
   }
@@ -99,9 +95,6 @@ class Antelope extends Animal {
         currentTile.setOrganism(null);
         return resolve();
       }
-      if (targetTile.organism.resistance) {
-        resolve();
-      }
       if (
         targetTile.organism.constructor === currentTile.organism.constructor
       ) {
@@ -112,26 +105,18 @@ class Antelope extends Animal {
         return resolve();
       }
       if (targetTile.organism.strength < currentTile.organism.strength) {
-        if (Math.random() < 0.5) {
-          this.xIndex = targetX;
-          this.yIndex = targetY;
-          this.board.removeOrganism(targetTile.organism);
-          targetTile.setOrganism(currentTile.organism);
-          currentTile.setOrganism(null);
-          return resolve();
-        } else {
-          return resolve();
-        }
-      }
-      if (Math.random() < 0.5) {
-        this.board.removeOrganism(currentTile.organism);
+        this.xIndex = targetX;
+        this.yIndex = targetY;
+        this.board.removeOrganism(targetTile.organism);
+        targetTile.setOrganism(currentTile.organism);
         currentTile.setOrganism(null);
         return resolve();
-      } else {
-        return resolve();
       }
+      this.board.removeOrganism(currentTile.organism);
+      currentTile.setOrganism(null);
+      return resolve();
     });
   }
 }
 
-export { Antelope };
+export { Animal };
