+++ src/organisms/animals/Antelope.js
@@ -1,6 +1,11 @@
-import { Organism } from './Organism';
+import { Animal } from './Animal';
+import { checkWithChance } from '../../utilities/checkWithChance';
 
-class Animal extends Organism {
+class Antelope extends Animal {
+  divClass = 'antelope-tile';
+  initiative = 4;
+  strength = 4;
+
   constructor(xIndex, yIndex, board) {
     super(xIndex, yIndex, board);
   }
@@ -9,55 +14,55 @@ class Animal extends Organism {
     if (randomNumber === 1) {
       return {
         xIndex: this.xIndex,
-        yIndex: this.yIndex - 1,
+        yIndex: this.yIndex - 2,
       };
     }
     if (randomNumber === 2) {
       return {
-        xIndex: this.xIndex + 1,
-        yIndex: this.yIndex - 1,
+        xIndex: this.xIndex + 2,
+        yIndex: this.yIndex - 2,
       };
     }
     if (randomNumber === 3) {
       return {
-        xIndex: this.xIndex + 1,
+        xIndex: this.xIndex + 2,
         yIndex: this.yIndex,
       };
     }
     if (randomNumber === 4) {
       return {
-        xIndex: this.xIndex + 1,
-        yIndex: this.yIndex + 1,
+        xIndex: this.xIndex + 2,
+        yIndex: this.yIndex + 2,
       };
     }
     if (randomNumber === 5) {
       return {
         xIndex: this.xIndex,
-        yIndex: this.yIndex + 1,
+        yIndex: this.yIndex + 2,
       };
     }
     if (randomNumber === 6) {
       return {
-        xIndex: this.xIndex - 1,
-        yIndex: this.yIndex + 1,
+        xIndex: this.xIndex - 2,
+        yIndex: this.yIndex + 2,
       };
     }
     if (randomNumber === 7) {
       return {
-        xIndex: this.xIndex - 1,
+        xIndex: this.xIndex - 2,
         yIndex: this.yIndex,
       };
     }
     if (randomNumber === 8) {
       return {
-        xIndex: this.xIndex - 1,
-        yIndex: this.yIndex - 1,
+        xIndex: this.xIndex - 2,
+        yIndex: this.yIndex - 2,
       };
     }
     if (randomNumber === 9) {
       return {
-        xIndex: this.xIndex - 1,
-        yIndex: this.yIndex - 1,
+        xIndex: this.xIndex - 2,
+        yIndex: this.yIndex - 2,
       };
     }
   }
@@ -95,6 +100,9 @@ class Animal extends Organism {
         currentTile.setOrganism(null);
         return resolve();
       }
+      if (targetTile.organism.resistance) {
+        resolve();
+      }
       if (
         targetTile.organism.constructor === currentTile.organism.constructor
       ) {
@@ -105,18 +113,26 @@ class Animal extends Organism {
         return resolve();
       }
       if (targetTile.organism.strength < currentTile.organism.strength) {
-        this.xIndex = targetX;
-        this.yIndex = targetY;
-        this.board.removeOrganism(targetTile.organism);
-        targetTile.setOrganism(currentTile.organism);
+        if (checkWithChance(0.5)) {
+          this.xIndex = targetX;
+          this.yIndex = targetY;
+          this.board.removeOrganism(targetTile.organism);
+          targetTile.setOrganism(currentTile.organism);
+          currentTile.setOrganism(null);
+          return resolve();
+        } else {
+          return resolve();
+        }
+      }
+      if (checkWithChance(0.5)) {
+        this.board.removeOrganism(currentTile.organism);
         currentTile.setOrganism(null);
         return resolve();
+      } else {
+        return resolve();
       }
-      this.board.removeOrganism(currentTile.organism);
-      currentTile.setOrganism(null);
-      return resolve();
     });
   }
 }
 
-export { Animal };
+export { Antelope };
