+++ src/organisms/animals/Fox.js
