+++ src/organisms/animals/Sheep.js
