+++ src/organisms/animals/Wolf.js
