+++ src/organisms/animals/deer.png
