+++ src/organisms/animals/face.png
