+++ src/organisms/animals/fox.png
