+++ src/organisms/animals/sheep.png
