+++ src/organisms/animals/wolf.png
