+++ src/organisms/plants/Berry.js
