+++ src/organisms/plants/Thistle.js
