+++ src/organisms/plants/grass3.png
