+++ src/organisms/plants/poisonberry.png
