+++ src/organisms/plants/sowthistle.png
