+++ src/utilities/checkWithChance.js
@@ -1,0 +1,5 @@
+function checkWithChance(numberToCheck) {
+  return Math.random() < numberToCheck;
+}
+
+export { checkWithChance };
